requirements.txt
torch
numpy
pyyaml
matplotlib
